Chirp/requirements.txt
google-cloud-speech>=2.33.0
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-storage>=2.18.0